# Conflicts:
#	CHANGELOGS.md
#	README.md
#	RELEASES.md
#	body/board/obj/body.bin
#	body/board/obj/body.bin.signed
#	body/board/obj/body.elf
#	body/board/obj/bootstub.body.bin
#	body/board/obj/bootstub.body.elf
#	body/board/obj/gitversion.h
#	cereal/SConscript
#	cereal/__init__.py
#	cereal/car.capnp
#	cereal/gen/cpp/car.capnp.c++
#	cereal/gen/cpp/car.capnp.h
#	cereal/gen/cpp/log.capnp.c++
#	cereal/gen/cpp/log.capnp.h
#	cereal/libcereal_shared.so
#	cereal/log.capnp
#	cereal/messaging/__init__.py
#	cereal/messaging/bridge
#	cereal/messaging/messaging.pxd
#	cereal/messaging/messaging_pyx.cpp
#	cereal/messaging/messaging_pyx.pyx
#	cereal/messaging/messaging_pyx.so
#	cereal/services.h
#	cereal/services.py
#	cereal/visionipc/__init__.py
#	cereal/visionipc/visionipc.pxd
#	cereal/visionipc/visionipc_pyx.cpp
#	cereal/visionipc/visionipc_pyx.pyx
#	cereal/visionipc/visionipc_pyx.so
#	common/clock.cpp
#	common/clock.so
#	common/kalman/simple_kalman_impl.cpp
#	common/kalman/simple_kalman_impl.so
#	common/logging_extra.py
#	common/params_pyx.cpp
#	common/params_pyx.so
#	common/transformations/transformations.cpp
#	common/transformations/transformations.so
#	common/util.h
#	common/version.h
#	docs/CARS.md
#	launch_chffrplus.sh
#	opendbc/can/common.pxd
#	opendbc/can/libdbc.so
#	opendbc/can/packer_pyx.cpp
#	opendbc/can/packer_pyx.pyx
#	opendbc/can/packer_pyx.so
#	opendbc/can/parser_pyx.cpp
#	opendbc/can/parser_pyx.pyx
#	opendbc/can/parser_pyx.so
#	opendbc/chrysler_pacifica_2017_hybrid_generated.dbc
#	opendbc/honda_civic_ex_2022_can_generated.dbc
#	opendbc/hyundai_canfd.dbc
#	opendbc/mazda_2017.dbc
#	opendbc/toyota_new_mc_pt_generated.dbc
#	opendbc/toyota_nodsu_pt_generated.dbc
#	opendbc/toyota_tnga_k_pt_generated.dbc
#	opendbc/vw_golf_mk4.dbc
#	panda/__init__.py
#	panda/board/obj/bootstub.panda.bin
#	panda/board/obj/bootstub.panda_h7.bin
#	panda/board/obj/panda.bin.signed
#	panda/board/obj/panda_h7.bin.signed
#	panda/python/__init__.py
#	panda/python/constants.py
#	panda/python/dfu.py
#	panda/python/spi.py
#	panda/python/usb.py
#	rednose/helpers/ekf_sym_pyx.cpp
#	rednose/helpers/ekf_sym_pyx.so
#	selfdrive/athena/athenad.py
#	selfdrive/boardd/boardd
#	selfdrive/boardd/boardd_api_impl.cpp
#	selfdrive/boardd/boardd_api_impl.so
#	selfdrive/boardd/panda.h
#	selfdrive/boardd/panda_comms.h
#	selfdrive/boardd/pandad.py
#	selfdrive/boardd/set_time.py
#	selfdrive/boardd/tests/test_boardd_loopback.py
#	selfdrive/car/__init__.py
#	selfdrive/car/body/carcontroller.py
#	selfdrive/car/chrysler/carcontroller.py
#	selfdrive/car/chrysler/carstate.py
#	selfdrive/car/chrysler/chryslercan.py
#	selfdrive/car/chrysler/interface.py
#	selfdrive/car/chrysler/values.py
#	selfdrive/car/docs_definitions.py
#	selfdrive/car/ford/carcontroller.py
#	selfdrive/car/ford/carstate.py
#	selfdrive/car/ford/fordcan.py
#	selfdrive/car/ford/interface.py
#	selfdrive/car/ford/radar_interface.py
#	selfdrive/car/ford/values.py
#	selfdrive/car/fw_query_definitions.py
#	selfdrive/car/fw_versions.py
#	selfdrive/car/gm/gmcan.py
#	selfdrive/car/gm/interface.py
#	selfdrive/car/gm/values.py
#	selfdrive/car/honda/carcontroller.py
#	selfdrive/car/honda/carstate.py
#	selfdrive/car/honda/interface.py
#	selfdrive/car/honda/values.py
#	selfdrive/car/hyundai/carcontroller.py
#	selfdrive/car/hyundai/carstate.py
#	selfdrive/car/hyundai/hyundaican.py
#	selfdrive/car/hyundai/hyundaicanfd.py
#	selfdrive/car/hyundai/interface.py
#	selfdrive/car/hyundai/radar_interface.py
#	selfdrive/car/hyundai/values.py
#	selfdrive/car/interfaces.py
#	selfdrive/car/isotp_parallel_query.py
#	selfdrive/car/mazda/carcontroller.py
#	selfdrive/car/mazda/carstate.py
#	selfdrive/car/mazda/interface.py
#	selfdrive/car/mazda/mazdacan.py
#	selfdrive/car/mazda/values.py
#	selfdrive/car/nissan/carcontroller.py
#	selfdrive/car/nissan/carstate.py
#	selfdrive/car/nissan/interface.py
#	selfdrive/car/nissan/nissancan.py
#	selfdrive/car/nissan/values.py
#	selfdrive/car/subaru/carcontroller.py
#	selfdrive/car/subaru/carstate.py
#	selfdrive/car/subaru/interface.py
#	selfdrive/car/subaru/subarucan.py
#	selfdrive/car/subaru/values.py
#	selfdrive/car/sunnypilot_carname.json
#	selfdrive/car/tesla/carcontroller.py
#	selfdrive/car/tesla/teslacan.py
#	selfdrive/car/tesla/values.py
#	selfdrive/car/tests/test_car_interfaces.py
#	selfdrive/car/toyota/interface.py
#	selfdrive/car/toyota/values.py
#	selfdrive/car/volkswagen/carcontroller.py
#	selfdrive/car/volkswagen/interface.py
#	selfdrive/car/volkswagen/mqbcan.py
#	selfdrive/car/volkswagen/pqcan.py
#	selfdrive/car/volkswagen/values.py
#	selfdrive/controls/controlsd.py
#	selfdrive/controls/lib/alerts_offroad.json
#	selfdrive/controls/lib/desire_helper.py
#	selfdrive/controls/lib/drive_helpers.py
#	selfdrive/controls/lib/events.py
#	selfdrive/controls/lib/lane_planner.py
#	selfdrive/controls/lib/lateral_mpc_lib/c_generated_code/libacados_ocp_solver_lat.so
#	selfdrive/controls/lib/lateral_planner.py
#	selfdrive/controls/lib/longitudinal_mpc_lib/c_generated_code/acados_ocp_solver_pyx.so
#	selfdrive/controls/lib/longitudinal_mpc_lib/c_generated_code/libacados_ocp_solver_long.so
#	selfdrive/controls/lib/longitudinal_mpc_lib/long_mpc.py
#	selfdrive/controls/lib/longitudinal_planner.py
#	selfdrive/controls/lib/radar_helpers.py
#	selfdrive/controls/lib/speed_limit_controller.py
#	selfdrive/controls/lib/vision_turn_controller.py
#	selfdrive/controls/plannerd.py
#	selfdrive/controls/radard.py
#	selfdrive/gpxd/gpxd.py
#	selfdrive/locationd/calibrationd.py
#	selfdrive/locationd/laikad.py
#	selfdrive/locationd/locationd
#	selfdrive/locationd/locationd.h
#	selfdrive/locationd/models/generated/car.cpp
#	selfdrive/locationd/models/generated/car.h
#	selfdrive/locationd/models/generated/gnss.cpp
#	selfdrive/locationd/models/generated/gnss.h
#	selfdrive/locationd/models/generated/libkf.so
#	selfdrive/locationd/models/generated/live.cpp
#	selfdrive/locationd/models/generated/live.h
#	selfdrive/locationd/paramsd.py
#	selfdrive/manager/custom_dep.py
#	selfdrive/manager/local_osm_install.py
#	selfdrive/manager/manager.py
#	selfdrive/manager/process.py
#	selfdrive/manager/process_config.py
#	selfdrive/mapd/config.py
#	selfdrive/mapd/mapd.py
#	selfdrive/modeld/_dmonitoringmodeld
#	selfdrive/modeld/_modeld
#	selfdrive/modeld/_navmodeld
#	selfdrive/modeld/models/driving.h
#	selfdrive/modeld/models/supercombo.thneed
#	selfdrive/navd/libmaprender.so
#	selfdrive/navd/map_renderer
#	selfdrive/sentry.py
#	selfdrive/test/helpers.py
#	selfdrive/test/test_onroad.py
#	selfdrive/thermald/thermald.py
#	selfdrive/ui/SConscript
#	selfdrive/ui/_ui
#	selfdrive/ui/qt/libpython_helpers.so
#	selfdrive/ui/qt/spinner
#	selfdrive/ui/qt/text
#	selfdrive/ui/soundd/_soundd
#	selfdrive/ui/translations/main_de.qm
#	selfdrive/ui/translations/main_de.ts
#	selfdrive/ui/translations/main_ja.qm
#	selfdrive/ui/translations/main_ja.ts
#	selfdrive/ui/translations/main_ko.qm
#	selfdrive/ui/translations/main_ko.ts
#	selfdrive/ui/translations/main_pt-BR.qm
#	selfdrive/ui/translations/main_pt-BR.ts
#	selfdrive/ui/translations/main_zh-CHS.qm
#	selfdrive/ui/translations/main_zh-CHS.ts
#	selfdrive/ui/translations/main_zh-CHT.qm
#	selfdrive/ui/translations/main_zh-CHT.ts
#	selfdrive/updated.py
#	system/camerad/camerad
#	system/clocksd/clocksd
#	system/hardware/hw.h
#	system/hardware/tici/hardware.h
#	system/logcatd/logcatd
#	system/loggerd/bootlog
#	system/loggerd/encoder/encoder.h
#	system/loggerd/encoder/ffmpeg_encoder.h
#	system/loggerd/encoder/v4l_encoder.h
#	system/loggerd/encoderd
#	system/loggerd/loggerd
#	system/loggerd/loggerd.h
#	system/logmessaged.py
#	system/proclogd/proclogd
#	system/sensord/_sensord
#	system/sensord/rawgps/rawgpsd.py
#	system/ubloxd/ubloxd
#	system/version.py
#	third_party/cluster/libfastcluster.so
#	tools/lib/helpers.py
#	tools/replay/replay.h

+++ selfdrive/manager/custom_dep.py
@@ -3,8 +3,11 @@ import os
 import sys
 import errno
 import shutil
+import tarfile
 import time
+import traceback
 from common.basedir import BASEDIR
+from common.text_window import TextWindow
 from urllib.request import urlopen
 from glob import glob
 import subprocess
@@ -21,6 +24,7 @@ MAX_BUILD_PROGRESS = 100
 TMP_DIR = '/data/tmp'
 THIRD_PARTY_DIR = '/data/openpilot/third_party/mapd'
 THIRD_PARTY_DIR_SP = '/data/third_party_community'
+PRELOADED_DEP_FILE = os.path.join(BASEDIR, "selfdrive/mapd/assets/mapd_deps.tar.xz")
 
 
 def wait_for_internet_connection(return_on_failure=False):
@@ -87,12 +91,55 @@ def install_dep(spinner):
 
 if __name__ == "__main__" and (OPSPLINE_SPEC is None or OVERPY_SPEC is None):
   spinner = Spinner()
-  if os.path.exists(THIRD_PARTY_DIR_SP):
-    spinner.update("Loading dependencies")
-    command = f'rm -rf {THIRD_PARTY_DIR}; cp -rf {THIRD_PARTY_DIR_SP} {THIRD_PARTY_DIR}'
+  preload_fault = False
+  try:
+    if os.path.exists(PRELOADED_DEP_FILE):
+      spinner.update("Loading preloaded dependencies")
+      try:
+        with tarfile.open(PRELOADED_DEP_FILE, "r:xz") as tar:
+          for member in tar.getmembers():
+            split_components = member.name.split('/')
+            if len(split_components) > 1:
+              member.name = '/'.join(split_components[1:])
+            tar.extract(member, path=THIRD_PARTY_DIR)
+        print(f"SP_LOG: Preloaded dependencies extracted to {THIRD_PARTY_DIR}")
+      except Exception as e:
+        preload_fault = True
+        command = f'rm -rf {THIRD_PARTY_DIR}'
+        process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
+        print(f"SP_LOG: An error occurred while extracting preloaded dependencies: {e}")
+        print(f"SP_LOG: Cleanup directory {e}")
+    if not os.path.exists(PRELOADED_DEP_FILE) or preload_fault:
+      if os.path.exists(THIRD_PARTY_DIR_SP):
+        try:
+          spinner.update("Loading cached dependencies")
+          command = f'rm -rf {THIRD_PARTY_DIR}; cp -rf {THIRD_PARTY_DIR_SP} {THIRD_PARTY_DIR}'
+          process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
+          print(f"SP_LOG: Removed directory {THIRD_PARTY_DIR}")
+          print(f"SP_LOG: Copied {THIRD_PARTY_DIR_SP} to {THIRD_PARTY_DIR}")
+        except Exception as e:
+          command = f'rm -rf {THIRD_PARTY_DIR}'
+          process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
+          print(f"SP_LOG: An error occurred while loading cached dependencies: {e}")
+          print(f"SP_LOG: Cleanup directory {e}")
+      else:
+        spinner.update("Waiting for internet")
+        try:
+          install_dep(spinner)
+        except Exception as e:
+          command = f'rm -rf {THIRD_PARTY_DIR}'
+          process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
+          print(f"SP_LOG: An error occurred while downloading dependencies: {e}")
+          print(f"SP_LOG: Cleanup directory {e}")
+  except Exception:
+    command = f'rm -rf {THIRD_PARTY_DIR}'
     process = subprocess.Popen(command, stdout=subprocess.PIPE, shell=True)
-    print(f"SP_LOG: Removed directory {THIRD_PARTY_DIR}")
-    print(f"SP_LOG: Copied {THIRD_PARTY_DIR_SP} to {THIRD_PARTY_DIR}")
-  else:
-    spinner.update("Waiting for internet")
-    install_dep(spinner)
+    import selfdrive.sentry as sentry
+    sentry.init(sentry.SentryProject.SELFDRIVE)
+    traceback.print_exc()
+    sentry.capture_exception()
+
+    error = traceback.format_exc(-3)
+    error = "Dependency Manager failed to start\n\n" + error
+    with TextWindow(error) as t:
+      t.wait_for_exit()
